Add `BaseActionGenerator`.
